backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.1,<4.0
pydantic-settings>=2.3,<3.0
aiosqlite>=0.20,<1.0
httpx>=0.27,<1.0
pytest>=8.2,<9.0
netmiko>=4.4,<5.0
slowapi>=0.1.9,<1.0
alembic>=1.13,<2.0
